Apply Markus review comment 9
